+++ SekibankiSpringMan/MyPluginInfo.cs
@@ -1,9 +1,0 @@
-namespace SekibankiSpringMan
-{
-    public static class MyPluginInfo
-    {
-        public const string PLUGIN_GUID = "com.bean.SekibankiSpringMan";
-        public const string PLUGIN_NAME = "[Touhou]SekibankiSpringMan(CoilHead)";
-        public const string PLUGIN_VERSION = "0.8.2";
-    }
-}