+++ TouhouEnemyModels/MyPluginInfo.cs
@@ -1,0 +1,9 @@
+namespace TouhouEnemyModels
+{
+    public static class MyPluginInfo
+    {
+        public const string PLUGIN_GUID = "com.bean.TouhouEnemyModels";
+        public const string PLUGIN_NAME = "TouhouEnemyModels";
+        public const string PLUGIN_VERSION = "0.9.0";
+    }
+}