+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# SekibankiSpringMan
+# TouhouEnemyModels
 ## Releases：
 ### Release v0.8.0
 1. Pre-release
@@ -6,3 +6,7 @@
 1. Fixed an issue that _CoilHeadStare_ think its eyes are on the top of its head.
 2. Fixed an issue that _VougeHeads_ makes the model not animate at all.
 [Unexpected Behavior with CoilHeadStare Installed +more](https://github.com/BoredBean/SekibankiSpringMan/issues/2)
+### Release v0.9.0
+1. Change Nutcracker to a Satori version.
+2. Add theme musics for Touhou characters.
+3. Optimized the shader and the texture.